+++ src/layer/application/api.ts
@@ -5,7 +5,7 @@ import { BroadcastChannel as WebStorage, localStorage, supportWebStorage } from 
 export * from '../domain/indexeddb/api';
 export * from '../domain/webstorage/api';
 
-export function storechannel<K extends string, V extends StoreChannelObject<K>>(name: string, config: StoreChannelConfig<K, V>): StoreChannel<K, V> {
+export function store<K extends string, V extends StoreChannelObject<K>>(name: string, config: StoreChannelConfig<K, V>): StoreChannel<K, V> {
   if (!supportWebStorage) throw new Error(`ClientChannel: Couldn't use WebStorage.`);
   const {
     schema,